# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.APIError:
    properties:
      detail:
        type: string
      error:
        type: string
    type: object
...
  handlers.APIHeartbeatRequest:
    properties:
      agentVersion:
        example: v2.27.0
        type: string
      labels:
        additionalProperties:
          type: string
        type: object
      osRelease:
        additionalProperties:
          type: string
        type: object
    type: object
  handlers.APIRegisterRequest:
    properties:
      agentVersion:
        example: v2.27.0
        type: string
      hostname:
        example: kairos-node-01
        type: string
      machineID:
        example: a1b2c3d4e5f6
        type: string
      registrationToken:
        example: 74a7e7452f9da2fd7739557f027bca0f
        type: string
    type: object
  handlers.APIRegisterResponse:
    properties:
      apiKey:
        example: 3db2c1e4f5a6b7c8d9e0f1a2b3c4d5e6
        type: string
      id:
        example: c8a4fb46-1836-4c70-97c8-29490ad110bc
        type: string
    type: object
...
  handlers.APISetGroupRequest:
    properties:
      groupID:
        example: b1db4f35-6195-410a-94f2-09e66e0a6a83
        type: string
    type: object
  handlers.APISetLabelsRequest:
    properties:
      labels:
        additionalProperties:
          type: string
        type: object
    type: object
...
  store.ManagedNode:
    properties:
      agentVersion:
        type: string
      createdAt:
        type: string
      group:
        $ref: '#/definitions/store.NodeGroup'
      groupID:
        type: string
      hostname:
        type: string
      id:
        type: string
      labels:
        additionalProperties:
          type: string
        type: object
      lastHeartbeat:
        type: string
      machineID:
        type: string
      osRelease:
        additionalProperties:
          type: string
        type: object
      phase:
        type: string
      updatedAt:
        type: string
    type: object
...
  store.NodeGroup:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
info:
  contact:
    name: Kairos authors
    url: https://github.com/kairos-io/AuroraBoot
  description: |-
    HTTP API for the AuroraBoot Kairos node manager.

    Three security schemes are used across the API:
    * AdminBearer — Authorization: Bearer <admin-password>. Human/UI/automation callers.
    * NodeAPIKey — Authorization: Bearer <api-key>. Registered Kairos nodes calling their own endpoints. The api-key is returned from POST /api/v1/nodes/register.
    * RegistrationToken — shared secret carried inside the body of POST /api/v1/nodes/register.

    Two WebSocket endpoints also exist but are not described in this spec:
    * GET /api/v1/ws — agent command channel (auth via ?token=<api-key>).
    * GET /api/v1/ws/ui — UI live update channel (auth via ?token=<admin-password>).
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  title: AuroraBoot API
  version: 0.1.0
paths:
...
  /api/v1/nodes:
    get:
      description: Returns every registered node. Optional group/label filters.
      parameters:
      - description: Filter by group ID
        in: query
        name: group
        type: string
      - description: Filter by a single key:value label pair
        in: query
        name: label
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.ManagedNode'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.APIError'
      security:
      - AdminBearer: []
      summary: List nodes
      tags:
      - Nodes
  /api/v1/nodes/{nodeID}:
    delete:
      parameters:
      - description: Node ID
        in: path
        name: nodeID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.APIError'
      security:
      - AdminBearer: []
      summary: Delete a node
      tags:
      - Nodes
    get:
      parameters:
      - description: Node ID
        in: path
        name: nodeID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.ManagedNode'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.APIError'
      security:
      - AdminBearer: []
      summary: Get a node
      tags:
      - Nodes
...
  /api/v1/nodes/{nodeID}/group:
    put:
      consumes:
      - application/json
      parameters:
      - description: Node ID
        in: path
        name: nodeID
        required: true
        type: string
      - description: Target group
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.APISetGroupRequest'
      responses:
        "200":
          description: OK
      security:
      - AdminBearer: []
      summary: Assign a node to a group
      tags:
      - Nodes
  /api/v1/nodes/{nodeID}/heartbeat:
    post:
      consumes:
      - application/json
      description: Transitions the node to Online and records the latest agent version
        and OS info.
      parameters:
      - description: Node ID
        in: path
        name: nodeID
        required: true
        type: string
      - description: Heartbeat payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.APIHeartbeatRequest'
      responses:
        "200":
          description: OK
      security:
      - NodeAPIKey: []
      summary: Agent heartbeat
      tags:
      - Agent
  /api/v1/nodes/{nodeID}/labels:
    put:
      consumes:
      - application/json
      parameters:
      - description: Node ID
        in: path
        name: nodeID
        required: true
        type: string
      - description: Labels payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.APISetLabelsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - AdminBearer: []
      summary: Replace a node's labels
      tags:
      - Nodes
  /api/v1/nodes/register:
    post:
      consumes:
      - application/json
      description: 'Idempotent by machineID: if a node with the same machineID already
        exists, the existing record is returned so the agent can resume with its persisted
        API key. Authenticated by the registrationToken inside the request body.'
      parameters:
      - description: Registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.APIRegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.APIRegisterResponse'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.APIRegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.APIError'
      summary: Register a node
      tags:
      - Agent bootstrap
...
securityDefinitions:
  AdminBearer:
    description: Supply as "Bearer <admin-password>".
    in: header
    name: Authorization
    type: apiKey
  NodeAPIKey:
    description: Supply as "Bearer <api-key>" returned from POST /api/v1/nodes/register.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"